requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.6
pillow==11.2.1
anthropic==0.49.0
fpdf2==2.8.3
python-dotenv==1.1.0
